fix(toast): fix toast display in app
